extact pageMiddleware() into its own file
